rename: PurchaseAmount, WinningLottoCounter, LottoGameRunner를 game 패키지로 이동
